requirements.txt
aiofiles==24.1.0
Pillow==10.4.0